Genralized keyword extraction for embedding models
