+++ py/genai_client/embedders/abstract_embedder.py
@@ -51,6 +51,62 @@ class AbstractEmbedder(ABC):
 
         return model_engine_response.to_dict()
 
+    def keyword_extraction(
+        self, input: List[str], percentile: int = 0, max_keywords: int = 12
+    ) -> List[str]:
+        from keybert import KeyBERT
+
+        kw_embedder = self.to_keybert_embedder()
+        kw_model = KeyBERT(model=kw_embedder)
+
+        list_of_chunks = input
+
+        keywords = self.get_text_keywords(
+            kw_model=kw_model,
+            list_of_chunks=list_of_chunks,
+            percentile=percentile,
+            max_keywords=max_keywords,
+        )
+
+        return keywords
+
+    def get_text_keywords(
+        self, kw_model, list_of_chunks: List[str], percentile: int, max_keywords: int
+    ) -> List[str]:
+
+        import numpy as np
+        from keyphrase_vectorizers import KeyphraseCountVectorizer
+
+        if len(list_of_chunks) == 1:
+            master_keywords_list = [
+                kw_model.extract_keywords(
+                    list_of_chunks,
+                    top_n=max_keywords,
+                    vectorizer=KeyphraseCountVectorizer(),
+                    use_mmr=True,
+                )
+            ]
+        else:
+            master_keywords_list = kw_model.extract_keywords(
+                list_of_chunks,
+                top_n=max_keywords,
+                vectorizer=KeyphraseCountVectorizer(),
+                use_mmr=True,
+            )
+
+        for i, keywords in enumerate(master_keywords_list):
+            if len(keywords) > 0:
+                keywords = keywords
+            else:
+                keywords = [("", 1.0)]
+
+            prob = [item[1] for item in keywords]
+            threshold = np.percentile(prob, percentile)
+            filtered_data = [word for word, score in keywords if score >= threshold]
+            master_keywords_list[i] = " ".join(filtered_data)
+
+        return master_keywords_list
+
     def to_keybert_embedder(self) -> BaseEmbedder:
 
         class CfgEmbedderBackend(BaseEmbedder):
