+++ py/genai_client/embedders/bedrock_embedder.py
@@ -105,63 +105,6 @@ class BedrockEmbedder(AbstractEmbedder):
             return json_obj
         else:
             raise ValueError(f"Unsupported model name: {self.model_name}")
-        
-    def keyword_extraction(
-        self, input: List[str], percentile: int = 0, max_keywords: int = 12
-    ) -> List[str]:
-        from keybert import KeyBERT
-
-        kw_embedder = self.to_keybert_embedder()
-        kw_model = KeyBERT(model=kw_embedder)
-        #kw_model.extract_keywords(docs=input)
-
-        list_of_chunks = input
-
-        keywords = BedrockEmbedder.get_text_keywords(
-            kw_model=kw_model,
-            list_of_chunks=list_of_chunks,
-            percentile=percentile,
-            max_keywords=max_keywords,
-        )
-
-        return keywords
-    
-    def get_text_keywords(
-        kw_model, list_of_chunks: List[str], percentile: int, max_keywords: int
-    ) -> List[str]:
-
-        import numpy as np
-        from keyphrase_vectorizers import KeyphraseCountVectorizer
-
-        if len(list_of_chunks) == 1:
-            master_keywords_list = [
-                kw_model.extract_keywords(
-                    list_of_chunks,
-                    top_n=max_keywords,
-                    vectorizer=KeyphraseCountVectorizer(),
-                    use_mmr=True,
-                )
-            ]
-        else:
-            master_keywords_list = kw_model.extract_keywords(
-                list_of_chunks,
-                top_n=max_keywords,
-                vectorizer=KeyphraseCountVectorizer(),
-                use_mmr=True,
-            )
-
-        for i, keywords in enumerate(master_keywords_list):
-            if len(keywords) > 0:
-                keywords = keywords
-            else:
-                keywords = [("", 1.0)]
-
-            prob = [item[1] for item in keywords]
-            threshold = np.percentile(prob, percentile)
-            filtered_data = [word for word, score in keywords if score >= threshold]
-            master_keywords_list[i] = " ".join(filtered_data)
-
-        return master_keywords_list
 
     def _get_tokenizer(self, init_args):
         return None
